parent 8e2985c4ed531edeef8cd6b5f44cbcd8a8609ad6 author AhmedBasem <demhamesab@gmail.com> 1687946371 +0300 committer Jusong Yu <jusong.yeu@gmail.com> 1692310651 +0200
Add PR preview action

Initialize React app.
Created the components and routes of the app.
Updated The webpage.yml with the deploy script.
Commit the plugins JSON file into git so that the React app can read the data from it.
Reused the current stylesheet and svg images

feat: sort plugins by commit count (#241)

Use github API to get the commits endpoint.
To get the number of commits in the last year we specify the period in the params of the request
and then extract the number of pages from the response header.

For plugins not hosted on github, we clone the repository and use git directly.

fix layout

deploy only when merged to master

Fix PR preview path issue

update

Display Entry points information & Add Sort feature (#245)

This commit introduces a new ranking of the yearly commits for sorting the plugins.

It also includes the implementation of showing the plugin descriptions to the plugin detail page.

Change checkbox icon (#248)

Using `CheckCircleIcon` from `@mui/icons-material/CheckCircle`

Add sidebar section (#247)

This addresses #246 and introduces a sidebar with plugin content outlines. To implement the sidebar, several considerations were taken into account:
* Keep the sidebar isolated from the header and footer
* Disable the sidebar for small screens (mobile phones)
* Make it in a fixed position, and allow for scrolling.

We used Materials UI `Drawer` component, it handles most of the above with less code.

Things are done in `App.jsx` and modulized to the have components in `components` folder.

The project folder structure
```
- public
- src
 -- assets
 -- Components
      ---MainIndex.jsx
      ---Details.jsx
      ---Sidebar.jsx
  --App.jsx
  --main.jsx
-index.html
```

Fix the website design on small screens (#253)